+++ tests/test_wisp.py
@@ -1,16 +1,16 @@
 import os
 import shutil
-import subprocess
 
 import numpy as np
+import pytest
 
 from wisp.run import run_wisp
 
 # Ensures we execute from file directory (for relative paths).
-os.chdir(os.path.dirname(os.path.realpath(__file__)))
+current_dir = os.path.dirname(os.path.abspath(__file__))
 
-FILE_DIR = "./files/"
-WRITING_DIR = "./tmp/"
+FILE_DIR = os.path.join(current_dir, "files/")
+WRITING_DIR = os.path.join(current_dir, "tmp/")
 
 
 def test_example():
@@ -49,7 +49,7 @@ def test_example():
         "spline_smoothness": 0.01,
         "vmd_resolution": 6,
         "output_directory": test_dir,
-        "logfile": open(os.path.join(test_dir, "log.txt"), "w"),
+        "logfile": open(os.path.join(test_dir, "log.txt"), "w", encoding="utf-8"),
         "user_specified_contact_map_filename": "",
         "user_specified_functionalized_matrix_filename": "",
         "simply_formatted_paths_filename": os.path.join(
@@ -114,12 +114,12 @@ def test_issue_1_serial():
         "shortest_path_opacity": 1.0,
         "shortest_path_r": 0.0,
         "shortest_path_radius": 1.0,
-        "sink_residues": ["A_HIE_840"],
-        "source_residues": ["A_ASN_692"],
+        "sink_residues": ["C_HIE_840"],
+        "source_residues": ["B_ASN_692"],
         "spline_smoothness": 0.01,
         "vmd_resolution": 6,
         "output_directory": test_dir,
-        "logfile": open(os.path.join(test_dir, "log.txt"), "w"),
+        "logfile": open(os.path.join(test_dir, "log.txt"), "w", encoding="utf-8"),
         "user_specified_contact_map_filename": "",
         "user_specified_functionalized_matrix_filename": "",
         "simply_formatted_paths_filename": os.path.join(
@@ -127,27 +127,5 @@ def test_issue_1_serial():
         ),
         "pdb_single_frame_filename": "",
     }
-    paths = run_wisp(config)
-    test_data = np.array([i[0] for i in paths], dtype=np.float64)
-
-    # ref_data = np.array(
-    #     [
-    #         1.1363589537262389,
-    #         2.027418997284591,
-    #         2.205901870787125,
-    #         2.307449850568712,
-    #         2.442623103273746,
-    #         2.448087831142992,
-    #         2.501932824731498,
-    #         2.5026984744745864,
-    #         2.621657347053458,
-    #         2.6370708377784413,
-    #         2.6804156982340315,
-    #         2.7056328296310643,
-    #         2.8932868716011986,
-    #         2.977212301921493,
-    #         3.0311375788312436,
-    #     ],
-    #     dtype=np.float64,
-    # )
-    # assert np.allclose(test_data, ref_data)
+    with pytest.raises(SystemExit) as e:
+        run_wisp(config)